+++ common/crc.h
@@ -19,6 +19,8 @@
 #pragma once
 #endif
 
+#include <stdint.h>
+
 // MD5 Hash
 typedef struct
 {
@@ -28,7 +30,7 @@ typedef struct
 } MD5Context_t;
 
 
-typedef unsigned long CRC32_t;
+typedef uint32_t CRC32_t;
 void CRC32_Init(CRC32_t *pulCRC);
 CRC32_t CRC32_Final(CRC32_t pulCRC);
 void CRC32_ProcessBuffer(CRC32_t *pulCRC, void *p, int len);
